requirements.txt
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
cryptography==36.0.2
Deprecated==1.2.13
Django==4.0.3
django-celery-beat==2.5.0
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-crontab==0.7.1
django-debug-toolbar==3.2.4
django-filter==23.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-oauth-toolkit==1.7.1
django-timezone-field==5.0
djangorestframework==3.13.1
drf-yasg==1.20.0
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.3
inflection==0.5.1
#install==1.3.5
itypes==1.2.0
Jinja2==3.1.1
jwcrypto==1.0
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.5
#mysqlclient==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.0.3
pyparsing==3.0.7
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.5.4
requests==2.27.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.6
wrapt==1.14.0
